+++ src/main/java/roomescape/config/CheckAdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.member.controller;
+package roomescape.config;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
