+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -4,8 +4,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.member.controller.CheckAdminInterceptor;
-import roomescape.member.controller.MemberArgumentResolver;
 import roomescape.member.service.AuthService;
 import java.util.List;
